+++ Makefile
@@ -3,3 +3,7 @@ importdb:
 
 run:
 	go run main.go
+
+all:
+	docker compose up -d
+	go run main.go
