+++ docker-compose.yml
@@ -19,14 +19,14 @@ services:
       - clickhouse
 
 
-#  prometheus:
-#    container_name: prometheus
-#    image: prom/prometheus:latest
-#    restart: always
-#    volumes:
-#      - ./prometheus.yaml:/etc/prometheus/prometheus.yml
-#    ports:
-#      - "9090:9090"
+  prometheus:
+    container_name: prometheus
+    image: prom/prometheus:latest
+    restart: always
+    volumes:
+      - ./prometheus.yaml:/etc/prometheus/prometheus.yml
+    ports:
+      - "9090:9090"
 
 
   clickhouse:
