+++ otel-collector-config.yaml
@@ -1,17 +1,3 @@
-exporters:
-  clickhouse:
-    endpoint: tcp://clickhouse:9000
-    database: otel
-    ttl_days: 0
-    timeout: 10s
-    sending_queue:
-      queue_size: 100
-    retry_on_failure:
-      enabled: true
-      initial_interval: 5s
-      max_interval: 30s
-      max_elapsed_time: 300s
-
 receivers:
   otlp:
     protocols:
@@ -20,8 +6,20 @@ receivers:
 
 processors:
   batch:
-    send_batch_size: 5
-    timeout: 5s
+
+exporters:
+  clickhouse:
+    endpoint: tcp://clickhouse:9000
+    database: otel
+    ttl_days: 0
+    timeout: 10s
+    sending_queue:
+      queue_size: 10000
+    retry_on_failure:
+      enabled: true
+      initial_interval: 5s
+      max_interval: 30s
+      max_elapsed_time: 300s
 
 service:
   extensions:
