+++ pkg/domain/BearTaskItem.go
@@ -3,7 +3,7 @@ package domain
 import "time"
 
 type BearTaskItem struct {
-	Date      time.Time `json:"date"`
-	Name      string    `json:"name"`
-	Completed bool      `json:"completed"`
+	Date      time.Time
+	Name      string
+	Completed bool
 }
